pyproject.toml
[project]
name = "metaquant"
version = "0.1.0"
description = "Modular quantitative research and stock screening framework"
authors = [{name = "MetaQuant Team"}]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "pydantic>=2.4",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "duckdb>=1.1.0",
    "rich>=13.7"
]

[project.scripts]
quantfw = "core.cli:app"

[tool.setuptools]
package-dir = {"" = "."}

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
